#352 Update to learning show and new partial
